refactor(sdk): split SDK types facade for GH727
