+++ specs/GH727/tasks.md
@@ -63,25 +63,32 @@ GH-727 / #727
 - [x] `SP727-T50` Owner: coordinator. Done when: URL construction moves to `src/core/providers/vertex_ai/client/url.rs` and health check moves to `src/core/providers/vertex_ai/client/health.rs`. Verify: `rg -n "build_url|get_publisher_for_model|check_health" src/core/providers/vertex_ai/client/*.rs`.
 - [x] `SP727-T51` Owner: coordinator. Done when: original inline Vertex AI client tests move to `src/core/providers/vertex_ai/client_tests.rs` without changing assertions. Verify: `rg -n "test_error_mapper_http_400|test_url_format_standard_location|test_vertex_ai_error_api_error" src/core/providers/vertex_ai/client_tests.rs`.
 - [x] `SP727-T52` Owner: verification owner. Done when: touched Vertex AI client files are below U-16's 800-line ceiling and focused tests pass. Verify: `wc -l src/core/providers/vertex_ai/client.rs src/core/providers/vertex_ai/client/*.rs src/core/providers/vertex_ai/client_tests.rs`; `cargo test core::providers::vertex_ai::client --lib --all-features`.
-- [ ] `SP727-T53` Owner: verification owner. Done when: formatting, all-features check, PR CI, and review-thread gate pass for the Vertex AI client tranche. Verify: `cargo fmt --all -- --check`; `cargo check --all-features --locked`; GitHub PR CI and review-thread query.
-- [ ] `SP727-T54` Owner: coordinator. Done when: after this tranche merges, the next #727 tranche is selected from the remaining queue with fresh line-count evidence. Verify: `rg --files -g '*.rs' src tests | xargs wc -l | awk '$1 > 800 && $2 != "total" { print $1 " " $2 }' | sort -nr`.
+- [x] `SP727-T53` Owner: verification owner. Done when: formatting, all-features check, PR CI, and review-thread gate pass for the Vertex AI client tranche. Verify: #814 PR body, green PR CI, and GraphQL review-thread query.
+- [x] `SP727-T54` Owner: coordinator. Done when: after this tranche merges, the next #727 tranche is selected from the remaining queue with fresh line-count evidence. Verify: at `origin/main@6e81da76d9a4`, `src/sdk/types.rs` is 1163 lines and 41 Rust files remain over the U-16 ceiling.
+- [x] `SP727-T55` Owner: coordinator. Done when: the SDK types tranche documents a facade-compatible split for message, tool, chat, and usage/cost DTOs. Verify: `git diff -- specs/GH727/product.md specs/GH727/tech.md`.
+- [x] `SP727-T56` Owner: coordinator. Done when: `src/sdk/types.rs` becomes a root facade that declares focused child modules and re-exports every original public type name. Verify: `sed -n '1,120p' src/sdk/types.rs`.
+- [x] `SP727-T57` Owner: coordinator. Done when: SDK production types are moved into `src/sdk/types/message.rs`, `tool.rs`, `chat.rs`, and `usage.rs` without field, enum, serde, or derive changes. Verify: `rg -n "pub (struct|enum)" src/sdk/types.rs src/sdk/types/*.rs`.
+- [x] `SP727-T58` Owner: coordinator. Done when: original inline SDK type tests are split by domain under `src/sdk/types_tests/` without changing assertions. Verify: `rg -n "test_role_variants|test_tool_call_creation|test_chat_request_creation|test_usage_default" src/sdk/types_tests/*.rs`.
+- [x] `SP727-T59` Owner: verification owner. Done when: all touched SDK type files are below U-16's 800-line ceiling and focused SDK type tests pass. Verify: `wc -l src/sdk/types.rs src/sdk/types/*.rs src/sdk/types_tests/*.rs`; `cargo test sdk::types --lib --all-features`.
+- [ ] `SP727-T60` Owner: verification owner. Done when: formatting, all-features check, PR CI, and review-thread gate pass for the SDK types tranche. Verify: `cargo fmt --all -- --check`; `cargo check --all-features --locked`; GitHub PR CI and review-thread query.
+- [ ] `SP727-T61` Owner: coordinator. Done when: after this tranche merges, the next #727 tranche is selected from the remaining queue with fresh line-count evidence. Verify: `git ls-files '*.rs' | xargs wc -l | awk '$1 > 800 && $2 != "total" { print $1 " " $2 }' | sort -nr`.
 
 ## 并行拆分
 
-This is a serial writable lane for the Vertex AI client file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
+This is a serial writable lane for the SDK types file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
 
 Writable ownership for this lane:
 
 - `specs/GH727/`
-- `src/core/providers/vertex_ai/client.rs`
-- `src/core/providers/vertex_ai/client/*.rs`
-- `src/core/providers/vertex_ai/client_tests.rs`
+- `src/sdk/types.rs`
+- `src/sdk/types/*.rs`
+- `src/sdk/types_tests/*.rs`
 
 ## 验证
 
 - SpecRail packet review.
 - `cargo fmt --all -- --check`
-- `cargo test core::providers::vertex_ai::client --lib --all-features`
+- `cargo test sdk::types --lib --all-features`
 - `cargo check --all-features --locked`
 - PR CI and GraphQL review-thread gate before merge.
